+++ app/src/main/java/com/cajetan/youtubeplus/fragments/StartFragment.kt
@@ -26,7 +26,7 @@ import org.jetbrains.anko.alert
 import org.jetbrains.anko.noButton
 import org.jetbrains.anko.yesButton
 
-class VideoListFragment : Fragment(), ContentListAdapter.ListItemClickListener,
+class StartFragment : Fragment(), ContentListAdapter.ListItemClickListener,
         YouTubeData.MostPopularListener, YouTubeData.VideoSearchListener,
         YouTubeData.UploadPlaylistListener {
 
@@ -55,7 +55,7 @@ class VideoListFragment : Fragment(), ContentListAdapter.ListItemClickListener,
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view = inflater.inflate(R.layout.fragment_video_list, container, false)
+        val view = inflater.inflate(R.layout.fragment_start, container, false)
 
         videoList = view.findViewById(R.id.videoList)
         searchProgressBarCentre = view.findViewById(R.id.searchProgressBarCentre)